refactor: feedComment - reader, writer
