+++ src/module/feed-comment/feed-comment.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { FeedCommentController } from './feed-comment.controller';
 import { FeedCommentService } from './feed-comment.service';
-import { FeedCommentRepository } from './repository/feed-comment.repository';
+import { FeedCommentReader } from './repository/feed-comment.reader';
+import { FeedCommentWriter } from './repository/feed-comment.writer';
 import { AwsModule } from '../aws/aws.module';
 
 @Module({
   imports: [AwsModule],
   controllers: [FeedCommentController],
-  providers: [FeedCommentService, FeedCommentRepository],
+  providers: [FeedCommentService, FeedCommentReader, FeedCommentWriter],
 })
 export class FeedCommentModule {}
